+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## main
+* Add `update` function and REPL command.
+
 ## 0.2.16 (2023-01-23)
 * Allow `JULIA_CONDAPKG_ENV` to specify the location of a shared Conda environment.
 * The PkgREPL now supports the prefixes `@` for versions and `#` for build string.